Abort Firefox setup if the latest version is already there.

+++ bin/firefox-latest-setup.sh
@@ -7,6 +7,10 @@ echo "LATEST=$LATEST"
 echo "LATEST_VER=$LATEST_VER"
 
 cd /usr/local
+if [ -d $LATEST_VER ] ; then
+    echo "The latest available version is already in place."
+    exit 1
+fi
 sudo rm firefox
 sudo tar xf $LATEST
 sudo mv firefox $LATEST_VER
